+++ src/components/ProductHome.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { Rating } from './Rating';
 
-const ProductHome = (props) => {
+export const ProductHome = (props) => {
     const {post} = props;
 
     return (
@@ -36,5 +36,3 @@ const ProductHome = (props) => {
         </>   
     )
 }
-
-export {ProductHome}
