+++ src/pages/Notfound.jsx
@@ -1,11 +1,9 @@
 import { Link } from 'react-router-dom';
 
-const Notfound = () => {
+export const Notfound = () => {
     return (
         <div>
             This page not found <Link to="/">Home</Link>
         </div>
     )
 }
-
-export {Notfound}
